rock-paper-scissors picture classification
First of all, I would like to thank Decoding Indonesia and DBS Foundation Coding Camp 2024 for the opportunity to join the academy program for free. I would like to introduce one of my final projects, which is image classification.

This project is a mini project or assignment that I worked on under the guidance of Decoding Indonesia. I am excited to share the result of this project: rock-paper-scissors image classification. Using the code that I developed, I managed to achieve an accuracy of around 90%.

I hope this project can give you an idea of my ability and dedication in the field of technology development. Thank you for this opportunity, and I am excited to share more of my experiences and future projects.